requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
certifi==2021.5.30
charset-normalizer==2.0.3
click==8.0.1
filelock==3.0.12
Flask==2.0.1
huggingface-hub==0.0.12
idna==3.2
importlib-metadata==4.6.1
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.0.1
MarkupSafe==2.0.1
numpy==1.21.0
packaging==21.0
Pillow==8.3.1
pyparsing==2.4.7
PyYAML==5.4.1
regex==2021.7.6
requests==2.26.0
sacremoses==0.0.45
six==1.16.0
tokenizers==0.10.3
torch==1.9.0+cpu
torchaudio==0.9.0
torchvision==0.10.0+cpu
tqdm==4.61.2
transformers==4.8.2
typing-extensions==3.10.0.0
urllib3==1.26.6
Werkzeug==2.0.1
zipp==3.5.0
